fix(Client): Correção de mensagens Flash

+++ src/main/java/com/dionialves/snapdogdelivery/client/ClientService.java
@@ -9,7 +9,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.dionialves.snapdogdelivery.client.dto.ClientDTO;
+import com.dionialves.snapdogdelivery.exception.BusinessException;
 import com.dionialves.snapdogdelivery.exception.NotFoundException;
+import com.dionialves.snapdogdelivery.order.OrderRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -18,6 +20,7 @@ import lombok.RequiredArgsConstructor;
 public class ClientService {
 
     private final ClientRepository clientRepository;
+    private final OrderRepository orderRepository;
 
     @Transactional(readOnly = true)
     public List<ClientDTO> search(String search) {
@@ -102,6 +105,10 @@ public class ClientService {
 
         if (!clientRepository.existsById(id)) {
             throw new NotFoundException("Client not found with ID: " + id);
+        }
+
+        if (orderRepository.existsByClientId(id)) {
+            throw new BusinessException("Customer cannot be deleted because it is associated with orders");
         }
         clientRepository.deleteById(id);
